code for handling message list seperated from code for handling message struct
